refactor: move error and loading component to tools repository
